refactor(abw): decompose listicle content generation
